feat: organize some styles
