+++ src/ui/molecules/form-credit-card/index.tsx
@@ -1,15 +1,16 @@
 "use client";
-import { Grid, TextField } from "@mui/material";
-import { useState } from "react";
-import { usePaymentInputs } from "react-payment-inputs";
+// @packages
 import images from "react-payment-inputs/images";
+import { Grid, TextField } from "@mui/material";
+import { usePaymentInputs } from "react-payment-inputs";
+import { useState } from "react";
 
 const FormCreditCard = () => {
   const {
-    getCardNumberProps,
-    getExpiryDateProps,
     getCVCProps,
     getCardImageProps,
+    getCardNumberProps,
+    getExpiryDateProps,
   } = usePaymentInputs();
   const [cardNumber, setCardNumber] = useState();
   const [experyDate, setExperyDate] = useState();
@@ -21,6 +22,10 @@ const FormCreditCard = () => {
     <Grid container gap={3}>
       <Grid item xs={12}>
         <TextField
+          fullWidth
+          label="Credit card number"
+          size="small"
+          variant="outlined"
           InputProps={{
             startAdornment: (
               <svg
@@ -35,64 +40,60 @@ const FormCreditCard = () => {
               onChange: handleChangeCardNumber,
             }),
           }}
-          fullWidth
-          size="small"
-          variant="outlined"
-          label="Credit card number"
         />
       </Grid>
       <Grid item container justifyContent="space-between">
         <Grid item xs={3}>
           <TextField
+            fullWidth
+            label="Expiry date"
+            size="small"
+            variant="outlined"
             inputProps={{
               value: experyDate,
               ...getExpiryDateProps({
                 onChange: (event: any) => setExperyDate(event.target.value),
               }),
             }}
-            size="small"
-            fullWidth
-            variant="outlined"
-            label="Expiry date"
           />
         </Grid>
         <Grid item xs={8}>
           <TextField
+            fullWidth
+            label="Security code (CVC)"
+            size="small"
+            variant="outlined"
             inputProps={{
               value: securityCode,
               ...getCVCProps({
                 onChange: (event: any) => setSecurityCode(event.target.value),
               }),
             }}
-            size="small"
-            fullWidth
-            variant="outlined"
-            label="Security code (CVC)"
           />
         </Grid>
       </Grid>
       <Grid item xs={12}>
         <TextField
-          size="small"
           fullWidth
-          variant="outlined"
           label="Cardholder's name"
+          size="small"
+          variant="outlined"
         />
       </Grid>
       <Grid item xs={12}>
         <TextField
-          size="small"
           fullWidth
-          variant="outlined"
           label="Cardholder ID"
+          size="small"
+          variant="outlined"
         />
       </Grid>
       <Grid item xs={12}>
         <TextField
-          size="small"
           fullWidth
-          variant="outlined"
           label="Number of payments"
+          size="small"
+          variant="outlined"
         />
       </Grid>
     </Grid>
