+++ src/ui/molecules/quantity-selector/index.module.css
@@ -1,0 +1,3 @@
+.iconButtonQuantity {
+  margin: "auto";
+}
