+++ src/ui/molecules/quantity-selector/index.tsx
@@ -1,10 +1,13 @@
 // @packages
-import IconButton from "@mui/material/IconButton";
 import ControlPointOutlinedIcon from "@mui/icons-material/ControlPointOutlined";
+import IconButton from "@mui/material/IconButton";
 import RemoveCircleOutlineOutlinedIcon from "@mui/icons-material/RemoveCircleOutlineOutlined";
-import { Grid } from "@mui/material";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
+import { Grid } from "@mui/material";
+
+// @styles
+import styles from "./index.module.css";
 
 type TQuantitySelector = {
   handleQuantity: (quantity: number) => void;
@@ -19,19 +22,19 @@ const QuantitySelector = ({ handleQuantity, quantity }: TQuantitySelector) => (
       </Typography>
     </Grid>
     <Grid item display="flex" alignContent="center">
-      <div style={{ margin: "auto" }}>
+      <div className={styles.iconButtonQuantity}>
         <IconButton size="small" onClick={() => handleQuantity(quantity - 1)}>
           <RemoveCircleOutlineOutlinedIcon fontSize="small" />
         </IconButton>
       </div>
       <TextField
         inputProps={{ min: 0, style: { textAlign: "center" } }}
+        onChange={(event) => handleQuantity(+event.target.value)}
         size="small"
         style={{ maxWidth: "100px", margin: "5px" }}
         value={quantity}
-        onChange={(event) => handleQuantity(+event.target.value)}
       />
-      <div style={{ margin: "auto" }}>
+      <div className={styles.iconButtonQuantity}>
         <IconButton onClick={() => handleQuantity(quantity + 1)} size="small">
           <ControlPointOutlinedIcon fontSize="small" />
         </IconButton>
