+++ src/ui/organisms/image-viewer/index.module.css
@@ -1,19 +1,17 @@
 .imageViewer {
   display: flex;
+  position: "relative";
 }
-
 
 .productImageWrapper {
-  padding: 12px;
-  display: flex;
-  justify-content: center;
   align-items: center;
   aspect-ratio: 4 / 3;
+  display: flex;
+  justify-content: center;
+  padding: 12px;
 }
 .productImage {
-  object-fit: cover;
-  max-width: 100%;
-  /* width: 100%;
-  height: 100%; */
   max-height: 100%;
+  max-width: 100%;
+  object-fit: cover;
 }
